fix link searching to detect all links in a given line
In the previous behavior, only the first link of a line was being detected, because the regex did not have the global flag, and it wasn't doing a `matchAll` check. 

In updated behavior, it does a match all, and iterates through the matches (i.e., the links), reducing the line at each step by updating each link.